Fix PubMed float64/float32 dtype mismatch

+++ scripts/data_adapters/pubmed_kg.py
@@ -157,7 +157,8 @@ def main():
     else:
         fingerprints = data["features"]
 
-    # Normalize
+    # Normalize and ensure float32
+    fingerprints = fingerprints.astype(np.float32)
     norms = np.linalg.norm(fingerprints, axis=1, keepdims=True)
     fingerprints = fingerprints / np.clip(norms, 1e-8, None)
